refactor(search): split searchController helpers into src/utils/search modules
Extract the ~1400 lines of pure helpers out of searchController.js into four
focused modules under src/utils/search, leaving the controller as a thin layer
with just the getUserLocation helper and the globalSearch / getAllUsersAndTeams
endpoints (2229 to 835 lines).

- searchParamParser.js: request query-string parsing (parseSearchParams)
- searchSqlBuilders.js: row normalizers plus WHERE/ORDER-BY/nearest-priority builders
- searchResultProcessing.js: tag/badge enrichment, viewer match scoring, boolean
  search-clause assembly, privacy sanitizers
- searchExecution.js: fetchOpenRoleSearchResults plus executeSearchQueries

Dependency direction is acyclic: controller to execution to
resultProcessing/sqlBuilders, with paramParser as a leaf. Public interface is
unchanged. The redundant distance-SQL wrapper methods (pure pass-throughs to
utils/searchQueryBuilder, unused outside this file) are dropped; searchExecution
calls the util directly. Verbatim function moves, no behavior change.

All 99 backend tests pass.
